backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0
numpy==1.26.2
openai==1.3.7
anthropic==0.7.8
beautifulsoup4==4.12.2
selenium==4.16.0
aiohttp==3.9.1
redis==5.0.1
